fix: docker file & entrypoint + added amqp config to telemetry service
