+++ Dockerfile
@@ -12,6 +12,7 @@ RUN rm -rf ./node-modules && \
     touch packages/server-core/.env && \
     touch packages/server-messenger/.env && \
     touch packages/server-analysis-manager/.env && \
+    touch packages/server-telemetry/.env && \
     touch packages/client-ui/.env
 
 COPY ./entrypoint.sh ./entrypoint.sh
@@ -24,4 +25,4 @@ HEALTHCHECK --interval=10s --timeout=5s --retries=5 \
     CMD wget --proxy off --no-verbose --tries=1 --spider http://127.0.0.1:3000/ || exit 1
 
 ENTRYPOINT ["/bin/sh", "./entrypoint.sh"]
-CMD ["cli", "start"]
+CMD ["core", "cli", "start"]
