+++ packages/server-kit/src/services/amqp/index.ts
@@ -6,3 +6,4 @@
  */
 
 export * from './module';
+export * from './singleton';
