+++ packages/server-telemetry/package.json
@@ -53,11 +53,11 @@
         "testcontainers": "^11.4.0"
     },
     "scripts": {
-        "dev": "ts-node src/cli/index.ts",
+        "cli-dev": "ts-node src/cli/index.ts",
         "build:types": "tsc --emitDeclarationOnly -p tsconfig.build.json",
         "build:js": "rollup -c",
         "build": "rimraf ./dist && cross-env NODE_ENV=production npm run build:js && npm run build:types",
-        "start": "node dist/cli/index.js",
+        "cli": "node dist/cli/index.js",
         "test": "cross-env NODE_ENV=test jest --config ./test/jest.config.js",
         "test:coverage": "npm run test -- --coverage"
     }
