+++ packages/server-telemetry/src/config/services/amqp.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2024.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+import { AmqpClient, setAmqpClientFactory } from '@privateaim/server-kit';
+import { useEnv } from '../env';
+
+export function configureAMQP() {
+    setAmqpClientFactory(() => new AmqpClient({
+        connectionOptions: useEnv('rabbitMqConnectionString'),
+    }));
+}
