+++ packages/server-telemetry/src/config/services/index.ts
@@ -5,6 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './amqp';
 export * from './authup';
 export * from './database';
 export * from './logger';
